Add files via upload
Added V3 of example files for chapters 4, 5, 6, 7, 8, 10, 11, 12, 13 updated to work with Qiskit v1.3, Qiskit runtime version: 0.34 and Python 3.12.
The setup_save_account.ipynb now gets the token from a file named "Token.txt" to be provided by the user. No need to update the code.
Added a header cell in every Jupyter notebook.
Added figure captions to facilitate cross referencing with the Paperback and PDF versions of the book.
Updated the code that prints the version of Qiskit, Qiskit runtime and Python.